DAAAARRRRRRRRRRRRRRRRRRRRRRRRRKKKKKKKKKKKKKKKKKKK!
Added some snippets from hanslman. I also noticed that
vim can be quite slow to load when I  first open certain python files,
I suspect it's because the code is horrendous so the linter is taking ages
to run it's report. If it gets to be an issue I suppose I could disable the
linter.

Or fix my code..